requirements.txt
streamlit>=1.28.0
gspread>=5.12.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pandas>=2.2.0
plotly>=5.18.0
Pillow>=10.1.0
qrcode[pil]>=7.4.2
python-barcode[images]>=0.15.1
pyyaml>=6.0.1
bcrypt>=4.1.1
streamlit-option-menu>=0.3.6
